move main module into separate location
