+++ src/client/src/main.ts
@@ -1,5 +1,11 @@
+import 'reflect-metadata';
+import 'core-js';
+import 'zone.js/dist/zone';
+import 'zone.js/dist/long-stack-trace-zone';
+
+
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
-import {AppModule} from './components/app.component';
+import {AppModule} from './modules/app.module';
 
 const platform = platformBrowserDynamic();
 platform.bootstrapModule(AppModule);
